refactor: moving transaction datasource to a service instead
